Effects defined with clones not directly
Insted of adding the effectIn/Out parameter in the object, we add this in a
clone of the object. Thus, we can have many effects and we can see them
more easily. In the future, the idea is to define the order by looking
where the clone is (with a layer of effects).